@@ -4,11 +4,11 @@ apply plugin: 'kotlin-android-extensions'
 
 android {
 
-    compileSdkVersion 27
+    compileSdkVersion 29
     defaultConfig {
         applicationId "dev.entao.QRApp"
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -54,14 +54,14 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation 'junit:junit:4.12'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
-    implementation 'dev.entao.kan:log:3.1.40'
-    implementation 'dev.entao.kan:json:3.1.40'
-    implementation 'dev.entao.kan:http:3.1.40'
-    implementation 'dev.entao.kan:appbase:3.1.40'
-    implementation 'dev.entao.kan:ui:3.1.40'
+    implementation 'dev.entao.kan:log:4.0.0'
+    implementation 'dev.entao.kan:json:4.0.0'
+    implementation 'dev.entao.kan:http:4.0.0'
+    implementation 'dev.entao.kan:appbase:4.0.0'
+    implementation 'dev.entao.kan:ui:4.0.0'
 
     implementation project(':qrlib')
 }
