@@ -6,15 +6,15 @@ apply plugin: 'maven'
 group = 'dev.entao.kan'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
 
     resourcePrefix 'zxing_'
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 1
-        versionName "3.1.40"
+        versionName "4.0.0"
     }
     buildTypes {
         release {
@@ -35,10 +35,11 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'com.google.zxing:core:3.3.3'
-    implementation 'dev.entao.kan:ui:3.1.40'
+
+    implementation 'dev.entao.kan:ui:4.0.0'
 }
 
 
@@ -54,11 +55,11 @@ artifacts {
 uploadArchives {
     repositories {
         mavenDeployer {
-            repository(url: 'http://app800.cn/maven/repository/releases/') {
-                authentication(userName: NEXUS_USERNAME, password: NEXUS_PASSWORD)
+            repository(url: 'http://maven.app800.cn/repository/public/') {
+                authentication(userName: ARCHIVA_USERNAME, password: ARCHIVA_PASSWORD)
             }
             pom.artifactId = "qr"
-            pom.groupId = project.group //'dev.entao'
+            pom.groupId = project.group
             pom.version = android.defaultConfig.versionName
             pom.packaging = 'aar'
 
