@@ -1,9 +1,9 @@
 package dev.entao.utilapp
 
 import android.os.Bundle
-import dev.entao.kan.base.ContainerActivity
+import dev.entao.kan.base.StackActivity
 
-class MainActivity : ContainerActivity() {
+class MainActivity : StackActivity() {
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
